Automate release procedure
Creates a manually callable GitHub Actions workflow that automates the
release procedure for the project. The workflow will determine the
correct version increment to apply and modify existing references to the
version in the code base. It will then create a release branch to build,
test, package, and and publish the release.